Fixed cast to float

+++ web/server.py
@@ -40,7 +40,7 @@ def get_db():
 @app.route('/add')
 def add_value():
     global notified
-    value = int(request.args.get('value', ''))
+    value = float(request.args.get('value', ''))
     db = get_db()
     reading = {"value": value,
                "date": datetime.datetime.utcnow()}
